% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-discovery.R
\name{analyze_existing_models_yaml}
\alias{analyze_existing_models_yaml}
\title{Update Database with YAML Analysis}
\usage{
analyze_existing_models_yaml(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\value{
Updated search state with analyzed covariates
}
\description{
Update Database with YAML Analysis
}
