% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-utils.R
\name{analyze_model_covariates_yaml}
\alias{analyze_model_covariates_yaml}
\title{Analyze Model Covariates from YAML}
\usage{
analyze_model_covariates_yaml(
  model_name,
  models_folder = "models",
  tags_list = list()
)
}
\arguments{
\item{model_name}{Character. Model name}

\item{models_folder}{Character. Models folder path}

\item{tags_list}{List. Available tags mapping}
}
\value{
Character vector of detected covariates
}
\description{
Analyze Model Covariates from YAML
}
