% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{analyze_model_from_logic}
\alias{analyze_model_from_logic}
\title{Analyze Model Using Logical Rules}
\usage{
analyze_model_from_logic(model_name)
}
\arguments{
\item{model_name}{Character. Model name}
}
\value{
Character vector indicating model type
}
\description{
Analyze Model Using Logical Rules
}
