% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clean_dir}
\alias{clean_dir}
\title{Remove all files starting with "WK_" from all model subdirectories}
\usage{
clean_dir(models_folder = "models")
}
\arguments{
\item{models_folder}{Character. Path to models directory (default: "models")}
}
\value{
Number of files deleted
}
\description{
Clean Directory - Remove NONMEM Working Files
}
