% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modification.R
\name{create_model_info_log}
\alias{create_model_info_log}
\title{Create detailed log file for model creation}
\usage{
create_model_info_log(
  search_state,
  model_name,
  parent_model,
  covariate_name,
  cov_info
)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. New model name}

\item{parent_model}{Character. Parent model name}

\item{covariate_name}{Character. Covariate added}

\item{cov_info}{Data.frame. Covariate information from search definition}
}
\value{
NULL (side effect: creates log file)
}
\description{
Creates a timestamped log file with model creation details
}
\details{
Create Model Info Log File
}
