% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R
\name{discover_existing_models}
\alias{discover_existing_models}
\title{Discover existing models and set up relationships}
\usage{
discover_existing_models(search_state)
}
\arguments{
\item{search_state}{List containing search state}
}
\value{
Updated search_state with discovered models
}
\description{
Scans the models folder and catalogs existing models
}
\details{
Discover Existing Models
}
