% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modification.R
\name{fix_theta_renumbering}
\alias{fix_theta_renumbering}
\title{Renumber THETA parameters after removing some}
\usage{
fix_theta_renumbering(modelcode, theta_numbers_to_remove, log_function)
}
\arguments{
\item{modelcode}{Character vector. Model file lines}

\item{theta_numbers_to_remove}{Numeric vector. THETA numbers that were removed}

\item{log_function}{Function. Logging function}
}
\value{
Character vector. Updated model code with renumbered THETAs
}
\description{
Adjusts THETA numbering throughout model file after removing THETAs
}
\details{
Fix THETA Renumbering
}
