% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{generate_phase}
\alias{generate_phase}
\title{Generate Phase (VECTORIZED VERSION)}
\usage{
generate_phase(step_number, action)
}
\arguments{
\item{step_number}{Integer vector of step numbers}

\item{action}{Character vector of action types}
}
\value{
Character vector of phases
}
\description{
Generate Phase (VECTORIZED VERSION)
}
