% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recovery.R
\name{generate_recovery_report}
\alias{generate_recovery_report}
\title{Generate Recovery Report}
\usage{
generate_recovery_report(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\value{
List with recovery statistics
}
\description{
Generate Recovery Report
}
