% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{generate_step_description}
\alias{generate_step_description}
\title{Generate Step Description (VECTORIZED VERSION)}
\usage{
generate_step_description(step_number, action, covariate_tested)
}
\arguments{
\item{step_number}{Integer vector of step numbers}

\item{action}{Character vector of action types}

\item{covariate_tested}{Character vector of covariate names}
}
\value{
Character vector of step descriptions
}
\description{
Generate Step Description (VECTORIZED VERSION)
}
