% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-selective-forward.R
\name{get_covariates_from_models}
\alias{get_covariates_from_models}
\title{Extract covariate names from specific models}
\usage{
get_covariates_from_models(search_state, model_names)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_names}{Character vector. Model names to extract covariates from}
}
\value{
Character vector of unique covariate tag names
}
\description{
Returns unique covariate names that were tested in the specified models
}
\details{
Get Covariates from Models
}
