% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-scm-core.R, R/07-scm-workflow.R
\name{get_dropped_covariates}
\alias{get_dropped_covariates}
\title{Get covariates that were tested but not selected in previous steps}
\usage{
get_dropped_covariates(search_state, model_name, tested_covariates)

get_dropped_covariates(search_state, model_name, tested_covariates)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Current model}

\item{tested_covariates}{Character. Previously tested covariate tags}

\item{current_model_id}{Character. Current best model}
}
\value{
Character vector of covariate tags that were dropped

Character vector of dropped covariate tags
}
\description{
Identifies covariates that were tested in earlier SCM steps
but not included in the current best model.
}
\details{
Get Dropped Covariates from Previous Steps
}
