% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-recovery.R
\name{get_excluded_covariates}
\alias{get_excluded_covariates}
\title{Get Excluded Covariates}
\usage{
get_excluded_covariates(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\value{
Character vector of excluded covariate names
}
\description{
Get Excluded Covariates
}
