% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-backward.R
\name{get_fixed_covariates}
\alias{get_fixed_covariates}
\title{Identify covariates with FIX status}
\usage{
get_fixed_covariates(search_state, covariate_names)
}
\arguments{
\item{search_state}{List containing search state}

\item{covariate_names}{Character vector. Covariate names to check}
}
\value{
Character vector of fixed covariate names
}
\description{
Checks both covariate_search data and model THETA parameters
to identify covariates that should not be removed
}
\details{
Get Fixed Covariates
}
