% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-models.R
\name{get_model_covariates}
\alias{get_model_covariates}
\title{Get Model Covariates}
\usage{
get_model_covariates(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name}
}
\value{
Character vector. Covariates in the model
}
\description{
Get Model Covariates
}
