% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{get_model_covariates_from_db}
\alias{get_model_covariates_from_db}
\title{Get all covariates in a model by tracing hierarchy}
\usage{
get_model_covariates_from_db(search_state, model_name)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. Model name}
}
\value{
Character vector. All covariate names in the model
}
\description{
Traces model hierarchy to get all covariates in a model
}
\details{
Get Model Covariates from Database (Cumulative)
}
