% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{get_model_covariates_from_files}
\alias{get_model_covariates_from_files}
\title{Extract covariates from model using BBR tags}
\usage{
get_model_covariates_from_files(search_state, model_name)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. Model name}
}
\value{
Character vector. Covariate names
}
\description{
Gets covariate information from BBR model tags
}
\details{
Get Model Covariates from Files
}
