% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{get_model_ofv_from_database}
\alias{get_model_ofv_from_database}
\title{Extract OFV from completed model database entry}
\usage{
get_model_ofv_from_database(search_state, model_name)
}
\arguments{
\item{search_state}{Search state object}

\item{model_name}{Model name}
}
\value{
Numeric OFV or NA if not found
}
\description{
Fast lookup of cached OFV value from search database
}
\details{
Get Model OFV from Database
}
