% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-io.R
\name{get_model_ofv_from_files}
\alias{get_model_ofv_from_files}
\title{Extract OFV from model output files}
\usage{
get_model_ofv_from_files(search_state, model_name)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. Model name}
}
\value{
Numeric. OFV value or NA
}
\description{
Extracts OFV value from NONMEM output files
}
\details{
Get Model OFV from Files
}
