% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-utils.R
\name{get_model_parent_yaml}
\alias{get_model_parent_yaml}
\title{Get Model Parent from YAML}
\usage{
get_model_parent_yaml(model_name, models_folder = "models")
}
\arguments{
\item{model_name}{Character. Model name}

\item{models_folder}{Character. Models folder path}
}
\value{
Character. Parent model name or NA
}
\description{
Get Model Parent from YAML
}
