% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R, R/04-models.R
\name{get_model_status}
\alias{get_model_status}
\title{Get current status of a model}
\usage{
get_model_status(search_state, model_name)

get_model_status(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name to check}
}
\value{
Character. Status: "completed", "failed", "in_progress", etc.

Character. Model status
}
\description{
Checks model completion status by examining output files
}
\details{
Get Model Status
}
