% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-algorithm.R
\name{get_remaining_covariates}
\alias{get_remaining_covariates}
\title{Get list of covariate tags that haven't been tested from base model}
\usage{
get_remaining_covariates(search_state, base_model_id, include_excluded = TRUE)
}
\arguments{
\item{search_state}{List containing covariate search state and configuration}

\item{base_model_id}{Character. Model to check current covariates against}

\item{include_excluded}{Logical. Whether to include excluded covariates (default: TRUE)}
}
\value{
Character vector of covariate tag names that can still be tested
}
\description{
Identifies which covariates from the search definition haven't
been added to the specified base model yet, with exclusion filtering.
}
\details{
Get Remaining Covariates for Testing
}
