% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{initialize_search_database_core}
\alias{initialize_search_database_core}
\title{Initialize Search Database}
\usage{
initialize_search_database_core(search_state)
}
\arguments{
\item{search_state}{List containing search state}
}
\value{
Updated search_state with an empty search database
}
\description{
Initialize Search Database
}
