% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-data.R
\name{load_search_data}
\alias{load_search_data}
\title{Load Search Data}
\usage{
load_search_data(data_file_path, covariate_search_path)
}
\arguments{
\item{data_file_path}{Character path to data file}

\item{covariate_search_path}{Character path to covariate search file}
}
\value{
List with data components
}
\description{
Load Search Data
}
