% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{load_search_state}
\alias{load_search_state}
\title{Load previously saved search state}
\usage{
load_search_state(filename)
}
\arguments{
\item{filename}{Character. Filename to load from}
}
\value{
List containing loaded search state
}
\description{
Loads a complete search state from an RDS file.
}
\details{
Load Search State from File
}
