% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R
\name{load_tags}
\alias{load_tags}
\title{Load covariate tags from YAML configuration}
\usage{
load_tags(search_state)
}
\arguments{
\item{search_state}{List containing search state}
}
\value{
Updated search_state with tags loaded
}
\description{
Loads the tags.yaml file containing covariate mappings
}
\details{
Load Tags from YAML File
}
