% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_imports.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{NULL-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side}

\item{y}{Right-hand side}
}
\value{
x if not NULL, otherwise y
}
\description{
NULL-coalescing operator
}
