% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-utilities.R
\name{read_model_file}
\alias{read_model_file}
\title{Read NONMEM control file with proper path handling}
\usage{
read_model_file(search_state, run_name, extensions = c(".ctl", ".mod"))
}
\arguments{
\item{search_state}{List containing search state}

\item{run_name}{Character. Model name}

\item{extensions}{Character vector. File extensions to try (default: c(".ctl", ".mod"))}
}
\value{
Character vector of model file lines with file_path attribute
}
\description{
Reads model control file (.ctl or .mod) and stores file path as attribute
}
\details{
Read Model File
}
