% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-utils.R
\name{read_model_yaml}
\alias{read_model_yaml}
\title{Read Model Information from YAML}
\usage{
read_model_yaml(model_name, models_folder = "models")
}
\arguments{
\item{model_name}{Character. Model name}

\item{models_folder}{Character. Models folder path}
}
\value{
List with model information
}
\description{
Read Model Information from YAML
}
