% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-exposure-forest.R
\name{theme_forest}
\alias{theme_forest}
\title{Forest-plot ggplot2 Theme}
\usage{
theme_forest(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size. Default 12.}

\item{base_family}{Base font family. Default "".}
}
\value{
A ggplot2 theme object to add to a plot.
}
\description{
A clean black-on-white \code{theme_bw()} variant used for the covariate
/ exposure forest plots (inward ticks, bordered panel, black text).
}
