% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{update_model_counter}
\alias{update_model_counter}
\title{Update model counter excluding retry models}
\usage{
update_model_counter(search_state)
}
\arguments{
\item{search_state}{List containing search state}
}
\value{
Updated search_state with model counter set
}
\description{
Sets the model counter based on existing models
}
\details{
Update Model Counter
}
