% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{validate_covariate_search_table}
\alias{validate_covariate_search_table}
\title{Validate covariate search table}
\usage{
validate_covariate_search_table(covariate_search, data_file)
}
\arguments{
\item{covariate_search}{data.frame with covariate search specification}

\item{data_file}{data.frame with analysis dataset}
}
\value{
validated covariate_search data.frame
}
\description{
Performs reusable checks on the covariate search table and returns the
validated table. Creates \code{cov_to_test} if it does not exist.
}
