% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-data.R
\name{validate_search_inputs}
\alias{validate_search_inputs}
\title{Validate Search Inputs}
\usage{
validate_search_inputs(base_model_path, data_file_path, covariate_search_path)
}
\arguments{
\item{base_model_path}{Character base model path}

\item{data_file_path}{Character data file path}

\item{covariate_search_path}{Character covariate search path}
}
\value{
Logical TRUE if valid
}
\description{
Validate Search Inputs
}
