% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-model-utilities.R
\name{view_search_database}
\alias{view_search_database}
\title{View the search database with model relationships}
\usage{
view_search_database(search_state, detailed = TRUE)
}
\arguments{
\item{search_state}{List containing search state}

\item{detailed}{Logical. Show detailed view (default: TRUE)}
}
\value{
Invisible search database
}
\description{
Displays a formatted view of the current search database
}
\details{
View Search Database
}
