% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calculate_covariate_df}
\alias{calculate_covariate_df}
\title{Determine degrees of freedom based on covariate type}
\usage{
calculate_covariate_df(covariate_name, covariate_search)
}
\arguments{
\item{covariate_name}{Character. Name of the covariate}

\item{covariate_search}{Data frame. Covariate search configuration}
}
\value{
Integer. Degrees of freedom (1 for continuous, n_levels-1 for categorical)
}
\description{
Calculates the degrees of freedom for a covariate based on whether
it is continuous or categorical. For categorical covariates, df equals the
number of levels minus 1.
}
\details{
Calculate Degrees of Freedom for Covariate
}
