% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{generate_changes_display}
\alias{generate_changes_display}
\title{Generate Changes Display (NEW FUNCTION FOR CHANGES COLUMN)}
\usage{
generate_changes_display(action, covariate_tested)
}
\arguments{
\item{action}{Character vector of action types}

\item{covariate_tested}{Character vector of covariate names}
}
\value{
Character vector of changes descriptions
}
\description{
Generate Changes Display (NEW FUNCTION FOR CHANGES COLUMN)
}
