% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{generate_step_display}
\alias{generate_step_display}
\title{Generate Step Display (NEW FUNCTION FOR STEP COLUMN)}
\usage{
generate_step_display(step_number, action)
}
\arguments{
\item{step_number}{Integer vector of step numbers}

\item{action}{Character vector of action types}
}
\value{
Character vector of step displays
}
\description{
Generate Step Display (NEW FUNCTION FOR STEP COLUMN)
}
