% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-scm-core.R, R/07-scm-steps.R
\name{get_dropped_covariates}
\alias{get_dropped_covariates}
\title{Get covariates that were tested but not selected in previous steps}
\usage{
get_dropped_covariates(search_state, current_model_id, tested_covariates)

get_dropped_covariates(search_state, current_model_id, tested_covariates)
}
\arguments{
\item{search_state}{List containing covariate search state and configuration}

\item{current_model_id}{Character. Current best model}

\item{tested_covariates}{Character vector. All covariate tags tested so far}
}
\value{
Character vector of covariate tags that were dropped

Character vector of covariate tags that were dropped
}
\description{
Identifies covariates that were tested in earlier SCM steps
but not included in the current best model.

Identifies covariates that were tested in earlier SCM steps
but not included in the current best model.
}
\details{
Get Dropped Covariates from Previous Steps

Get Dropped Covariates
}
