% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-algorithm.R
\name{get_excluded_covariates}
\alias{get_excluded_covariates}
\title{Get list of covariates excluded from current step}
\usage{
get_excluded_covariates(search_state)
}
\arguments{
\item{search_state}{List containing covariate search state and configuration}
}
\value{
Character vector of excluded covariate names
}
\description{
Returns covariates that have been excluded due to estimation issues
}
\details{
Get Excluded Covariates
}
