% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{get_model_covariates}
\alias{get_model_covariates}
\title{Get list of covariates in a model from database}
\usage{
get_model_covariates(search_state, model_name)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. Model name}
}
\value{
Character vector. Covariate names in the model
}
\description{
Returns covariates currently in the specified model
}
\details{
Get Model Covariates
}
