% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R, R/04-models.R
\name{get_model_covariates}
\alias{get_model_covariates}
\title{Get list of covariates in a model}
\usage{
get_model_covariates(search_state, model_name)

get_model_covariates(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name}
}
\value{
Character vector. Covariate tag names

Character vector. Covariates in the model
}
\description{
Extracts covariate tags from model using BBR tags
}
\details{
Get Model Covariates
}
