% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R, R/04-models.R
\name{get_model_ofv}
\alias{get_model_ofv}
\title{Extract OFV from completed model}
\usage{
get_model_ofv(search_state, model_name)

get_model_ofv(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name}
}
\value{
Numeric. OFV value or NA if not available

Numeric. Model OFV or NA
}
\description{
Extracts the objective function value from model output
}
\details{
Get Model OFV
}
