% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-models.R
\name{get_model_ofv}
\alias{get_model_ofv}
\title{Get Model OFV}
\usage{
get_model_ofv(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name}
}
\value{
Numeric. Model OFV or NA
}
\description{
Get Model OFV
}
