% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R
\name{get_model_status}
\alias{get_model_status}
\title{Get current status of a model from database}
\usage{
get_model_status(search_state, model_name)
}
\arguments{
\item{search_state}{List containing search state}

\item{model_name}{Character. Model name to check}
}
\value{
Character. Status: "completed", "failed", "in_progress", etc.
}
\description{
Checks model completion status from search database
}
\details{
Get Model Status
}
