% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-models.R
\name{get_model_status}
\alias{get_model_status}
\title{Get Model Status}
\usage{
get_model_status(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name to check}
}
\value{
Character. Model status
}
\description{
Get Model Status
}
