% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-file-io.R
\name{get_model_status_from_files}
\alias{get_model_status_from_files}
\title{Get Model Status from Files}
\usage{
get_model_status_from_files(model_path)
}
\arguments{
\item{model_path}{Character. Path to model directory}
}
\value{
Character. Overall model status
}
\description{
Determines overall model status from NONMEM output files
}
