% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-scm-core.R, R/07-scm-steps.R
\name{get_remaining_covariates}
\alias{get_remaining_covariates}
\title{Get list of covariate tags that haven't been tested from base model}
\usage{
get_remaining_covariates(search_state, base_model_id)

get_remaining_covariates(search_state, base_model_id)
}
\arguments{
\item{search_state}{List containing covariate search state and configuration}

\item{base_model_id}{Character. Model to check current covariates against}
}
\value{
Character vector of covariate tag names that can still be tested

Character vector of covariate tag names that can still be tested
}
\description{
Identifies which covariates from the search definition haven't
been added to the specified base model yet.

Identifies which covariates from the search definition haven't
been added to the specified base model yet.
}
\details{
Get Remaining Covariates for Testing

Get Remaining Covariates
}
