% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-selective-forward.R
\name{get_significant_models_from_step}
\alias{get_significant_models_from_step}
\title{Extract models that showed significant improvement in a specific step}
\usage{
get_significant_models_from_step(
  search_state,
  step_number,
  ofv_threshold,
  rse_threshold = NULL
)
}
\arguments{
\item{search_state}{List containing search state}

\item{step_number}{Integer. Step number to check}

\item{ofv_threshold}{Numeric. OFV threshold for significance}

\item{rse_threshold}{Numeeric. RSE threshold for significance}
}
\value{
Character vector of significant model names
}
\description{
Returns model names from a step that have ΔOFV above threshold
}
\details{
Get Significant Models from Step
}
