% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-initialize.R
\name{initialize_covariate_search}
\alias{initialize_covariate_search}
\title{Initialize CovariateSearcher State}
\usage{
initialize_covariate_search(
  base_model_path,
  data_file_path,
  covariate_search_path,
  models_folder = "models",
  threads = 4
)
}
\arguments{
\item{base_model_path}{Character. Path to base NONMEM model}

\item{data_file_path}{Character. Path to dataset file}

\item{covariate_search_path}{Character. Path to covariate search definition}

\item{models_folder}{Character. Directory for NONMEM models}

\item{threads}{Integer. Number of threads for parallel execution}
}
\value{
List containing initialized search state
}
\description{
Initialize CovariateSearcher State
}
