% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{initialize_search_database}
\alias{initialize_search_database}
\title{Initialize Search Database}
\usage{
initialize_search_database(models_folder)
}
\arguments{
\item{models_folder}{Character. Path to models directory}
}
\value{
List with database and counter
}
\description{
Initialize Search Database
}
