% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-results.R
\name{print_scm_results_table}
\alias{print_scm_results_table}
\title{Display SCM results table with formatting}
\usage{
print_scm_results_table(search_state)
}
\arguments{
\item{search_state}{List containing search state}
}
\description{
Print SCM Results Table
}
