% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-file-io.R
\name{read_nonmem_lst}
\alias{read_nonmem_lst}
\title{Read NONMEM LST File Status}
\usage{
read_nonmem_lst(model_path)
}
\arguments{
\item{model_path}{Character. Path to model directory or lst file}
}
\value{
List with status information
}
\description{
Reads NONMEM .lst file to determine model status and issues
}
