% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scm-execution.R
\name{run_univariate_step}
\alias{run_univariate_step}
\title{Run Univariate Step (SIMPLIFIED DATABASE VERSION)}
\usage{
run_univariate_step(search_state, base_model_id, covariates_to_test, step_name)
}
\arguments{
\item{search_state}{List containing covariate search state and configuration}

\item{base_model_id}{Character. Base model to test from}

\item{covariates_to_test}{Character vector. Covariate tags to test}

\item{step_name}{Character. Description for this step}
}
\value{
List with created model information and updated search_state
}
\description{
Run Univariate Step (SIMPLIFIED DATABASE VERSION)
}
