% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-management.R
\name{save_search_state}
\alias{save_search_state}
\title{Save Search State to Models Folder}
\usage{
save_search_state(search_state, filename)
}
\arguments{
\item{search_state}{Search state object to save}

\item{filename}{Name of the file (without path)}
}
\value{
Invisible path to saved file
}
\description{
Wrapper function that saves search state files to the models folder
instead of project root, keeping all SCM artifacts together
}
