% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{update_all_model_statuses}
\alias{update_all_model_statuses}
\title{Update All Model Statuses}
\usage{
update_all_model_statuses(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\value{
List with updated search_state
}
\description{
Updates all models in database from NONMEM output files
}
