% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-core-functions.R, R/05-validation.R
\name{update_all_model_statuses}
\alias{update_all_model_statuses}
\title{Update status for all models in database}
\usage{
update_all_model_statuses(search_state)

update_all_model_statuses(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\value{
Updated search_state with refreshed model statuses

Updated search state
}
\description{
Refreshes status information for all tracked models

Updates all models in database from NONMEM output files
}
\details{
Update All Model Statuses
}
