% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{update_model_status_from_files}
\alias{update_model_status_from_files}
\title{Update Model Status in Database}
\usage{
update_model_status_from_files(search_state, model_name)
}
\arguments{
\item{search_state}{List. Current search state}

\item{model_name}{Character. Model name to update}
}
\value{
List with updated search_state
}
\description{
Updates search database with results from NONMEM output
}
