% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-database.R
\name{view_comprehensive_table}
\alias{view_comprehensive_table}
\title{View Model Summary Table}
\usage{
view_comprehensive_table(search_state)
}
\arguments{
\item{search_state}{List. Current search state}
}
\description{
View Model Summary Table
}
